requirements.txt
flask==2.2.3
flask_sqlalchemy==3.0.3
flask_login==0.6.2
flask_wtf==1.1.1
flask_mail==0.9.1
email_validator==2.0.0.post2
werkzeug==2.2.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
wtforms==3.0.1
sqlalchemy==2.0.16
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.3
gunicorn==20.1.0
blinker==1.6.2
dataclasses==0.8; python_version < '3.7'
